refactor(web): 페이지 전수조사 기반 UI/UX 중복 제거·통합 1차
전 페이지(워크스페이스 17·설정 9탭·관리자 11) 기능 인벤토리에서 확인된 중복·죽은 UI 를
정리한다. 라우트는 유지(딥링크·E2E 호환), 승인 창구 단일 원칙(/approvals) 완성.

승인 창구 단일화
- /agent-tasks 인라인 ApprovalsPanel 제거 — /approvals TaskApprovals 와 동일 엔드포인트의
  이중 노출. 그 패널에만 있던 "이 작업 자동 승인"은 TaskApprovals 로 이관
- /custom-agents "Draft 검토" 탭 → /approvals 4번째 섹션(CustomAgentDrafts). Git URL
  가져오기 draft 만 승인 대상이라는 사실을 hint 로 명시. 사이드바 승인 뱃지 합산에 포함

설정 중복·죽은 컨트롤
- 일반 탭 "대화 기록 저장" 제거(개인정보 탭과 동일 state·키)
- 인터페이스 테마 Select → next-themes 직접 구동(preferences.theme 은 저장만 되고 소비처 0)
- 알림 "이메일 알림" 제거(preferences.emailAlerts 는 저장만 되고 발송 로직이 읽지 않음)

프로덕션에 실렌더되던 하드코딩 목업 제거
- 커넥터 buildMockServers(5종) · MCP 카탈로그 buildMockCatalog(+목업 전용 도구 수 배지)
  · 커스텀 에이전트 AGENTS_FALLBACK_META · 스킬 라이브러리 buildFallback
  · 에이전트 학습 ITEMS_FALLBACK(실 API 없음 → 빈 상태)
- skill-library 미사용 DiffView import·DraftsResponse 타입 제거
- 고아 i18n 키 11건 4개 locale 동시 삭제
