+++ apps/web/app/(workspace)/agent-tasks/page.tsx
@@ -502,141 +502,6 @@ function TaskDetailModal({
 
 /* ── 메인 페이지 ──────────────────────────────────────────── */
 /* ── 승인 대기 패널 (HITL 게이트) ──────────────────────────── */
-interface PendingApproval {
-  approvalId: string;
-  taskId: string;
-  toolName: string;
-  args: Record<string, unknown>;
-}
-type ApprovalsResponse = ApiSuccess<{ pending: PendingApproval[] }>;
-
-function ApprovalsPanel() {
-  const t = useTranslations("agentTasks");
-  const [pending, setPending] = useState<PendingApproval[]>([]);
-  const [busy, setBusy] = useState<string | null>(null);
-  const [answers, setAnswers] = useState<Record<string, string>>({});
-
-  const load = useCallback(async () => {
-    try {
-      const res = await ApiClient.get<ApprovalsResponse>("/api/agent-tasks/approvals/pending");
-      setPending(res?.data?.pending ?? []);
-    } catch {
-      // 401·네트워크: 빈 목록 유지
-    }
-  }, []);
-
-  useEffect(() => {
-    load();
-    const id = setInterval(load, 4000);
-    return () => clearInterval(id);
-  }, [load]);
-
-  async function run(approvalId: string, fn: () => Promise<void>) {
-    setBusy(approvalId);
-    try {
-      await fn();
-      await load();
-    } catch (err) {
-      alert(t("processFailed", { message: err instanceof Error ? err.message : t("error") }));
-    } finally {
-      setBusy(null);
-    }
-  }
-
-  const decide = (approvalId: string, decision: "approve" | "reject") =>
-    run(approvalId, () => ApiClient.post(`/api/agent-tasks/approvals/${approvalId}/${decision}`, {}));
-  const answer = (approvalId: string) =>
-    run(approvalId, () => ApiClient.post(`/api/agent-tasks/approvals/${approvalId}/answer`, { text: answers[approvalId] ?? "" }));
-  // task 자동승인(4-2) — 이후 이 작업의 도구 호출은 승인 없이 진행(ask_human 제외).
-  const autoApprove = (taskId: string, approvalId: string) =>
-    run(approvalId, () => ApiClient.post(`/api/agent-tasks/${taskId}/approvals/auto-approve`, {}));
-
-  if (pending.length === 0) return null;
-  const firstToolApproval = pending.find((p) => p.toolName !== "ask_human");
-
-  return (
-    <Card className="mb-4 p-4">
-      <div className="mb-3 flex items-center justify-between gap-2">
-        <p className="text-sm font-medium text-fg-2">
-          {t("approvalsTitle", { count: pending.length })}
-        </p>
-        {firstToolApproval && (
-          <Button
-            size="sm"
-            variant="outline"
-            disabled={busy !== null}
-            title={t("autoApproveHint")}
-            onClick={() => autoApprove(firstToolApproval.taskId, firstToolApproval.approvalId)}
-          >
-            {t("autoApprove")}
-          </Button>
-        )}
-      </div>
-      <div className="space-y-2">
-        {pending.map((p) => {
-          // ask_human 은 도구 승인이 아니라 사용자 질문 — 자유텍스트 답변 채널을 렌더.
-          if (p.toolName === "ask_human") {
-            const question = typeof p.args?.question === "string" ? p.args.question : "";
-            const text = answers[p.approvalId] ?? "";
-            return (
-              <div key={p.approvalId} className="space-y-2 rounded-md border border-line bg-bg-1 p-2">
-                <p className="text-xs font-semibold text-fg-2">{t("question")}</p>
-                {question && <p className="break-words text-xs text-fg-1">{question}</p>}
-                <textarea
-                  value={text}
-                  onChange={(e) => setAnswers((prev) => ({ ...prev, [p.approvalId]: e.target.value }))}
-                  placeholder={t("answerPlaceholder")}
-                  rows={2}
-                  disabled={busy === p.approvalId}
-                  className="w-full resize-y rounded-md border border-line bg-bg-2 p-1.5 text-xs text-fg-1 disabled:opacity-50"
-                />
-                <div className="flex justify-end gap-1">
-                  <Button size="sm" variant="outline" disabled={busy === p.approvalId} onClick={() => decide(p.approvalId, "reject")}>
-                    {t("skip")}
-                  </Button>
-                  <Button size="sm" disabled={busy === p.approvalId || !text.trim()} onClick={() => answer(p.approvalId)}>
-                    {t("sendAnswer")}
-                  </Button>
-                </div>
-              </div>
-            );
-          }
-          return (
-            <div
-              key={p.approvalId}
-              className="flex items-center justify-between gap-3 rounded-md border border-line bg-bg-1 p-2"
-            >
-              <div className="min-w-0">
-                <span className="font-mono text-xs text-fg-2">{p.toolName}</span>
-                <span className="ml-2 break-all text-xs text-muted">
-                  {JSON.stringify(p.args).slice(0, 100)}
-                </span>
-              </div>
-              <div className="flex shrink-0 gap-1">
-                <Button
-                  size="sm"
-                  variant="outline"
-                  disabled={busy === p.approvalId}
-                  onClick={() => decide(p.approvalId, "reject")}
-                >
-                  {t("reject")}
-                </Button>
-                <Button
-                  size="sm"
-                  disabled={busy === p.approvalId}
-                  onClick={() => decide(p.approvalId, "approve")}
-                >
-                  {t("approve")}
-                </Button>
-              </div>
-            </div>
-          );
-        })}
-      </div>
-    </Card>
-  );
-}
-
 /* ── 스케줄(반복 트리거) 패널 (Phase 3-A) ───────────────────── */
 interface ApiSchedule {
   id: string;
@@ -1040,7 +905,6 @@ export default function AgentTasksPage() {
             </Button>
           </div>
         </Card>
-        <ApprovalsPanel />
         <SchedulesPanel />
         <TemplatesPanel />
         {loading ? (
